pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = "Тестовое задание от DOCLUB-TV"
authors = ["Damir"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
pydantic = "1.10.8"
django-database-url = "^1.0.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
black = "^24.10.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"
drf-nested-routers = "^0.94.1"
pillow = "^11.0.0"
gunicorn = "^23.0.0"
django-debug-toolbar = "^4.4.6"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
